Os métodos foram atualizados
